+++ server/src/routes/template/mail_template.html
@@ -34,9 +34,7 @@
 <body>
   <h4>아래 버튼을 클릭하시면 돼지짱 가입이 승인됩니다.</h4>
   <br />
-  <a href="{replacetext}">
-    <button>가입확인</button>
-  </a>
+    <h4>{replacetext}</h4>
   
 </body>
 </html>
